fixed node setup and initialization

+++ sequencer/common/pooll2tx.go
@@ -123,7 +123,7 @@ func (tx *PoolL2Tx) SetID() error {
 // [ 16 bits ] chainId // 2 bytes
 // [ 32 bits ] signatureConstant // 4 bytes
 // Total bits compressed data:  225 bits // 29 bytes in *big.Int representation
-func (tx *PoolL2Tx) TxCompressedData(chainID uint16) (*big.Int, error) {
+func (tx *PoolL2Tx) TxCompressedData(chainID uint64) (*big.Int, error) {
 	var b [29]byte
 
 	toBJJSign := byte(0)
@@ -150,7 +150,7 @@ func (tx *PoolL2Tx) TxCompressedData(chainID uint16) (*big.Int, error) {
 		return nil, Wrap(err)
 	}
 	copy(b[17:23], fromIdxBytes[:])
-	binary.BigEndian.PutUint16(b[23:25], chainID)
+	binary.BigEndian.PutUint64(b[23:25], chainID)
 	copy(b[25:29], SignatureConstantBytes[:])
 
 	bi := new(big.Int).SetBytes(b[:])
@@ -311,7 +311,7 @@ func (tx *PoolL2Tx) RqTxCompressedDataV2() (*big.Int, error) {
 
 // HashToSign returns the computed Poseidon hash from the *PoolL2Tx that will
 // be signed by the sender.
-func (tx *PoolL2Tx) HashToSign(chainID uint16) (*big.Int, error) {
+func (tx *PoolL2Tx) HashToSign(chainID uint64) (*big.Int, error) {
 	toCompressedData, err := tx.TxCompressedData(chainID)
 	if err != nil {
 		return nil, Wrap(err)
